+++ lib/widgets/animation/animation_child/fade/fade_transition_view.dart
@@ -1,27 +1,27 @@
 // © COPYRIGHT 2022 APPDADDY SOFTWARE SOLUTIONS INC. ALL RIGHTS RESERVED.
 import 'package:flutter/material.dart';
 import 'package:fml/widgets/animation/animation_helper.dart';
-import 'package:fml/widgets/animation/animation_transition/size/size_transition_model.dart'
+import 'package:fml/widgets/animation/animation_child/fade/fade_transition_model.dart'
     as MODEL;
 import 'package:fml/widgets/widget/widget_model.dart';
 
 /// Animation View
 ///
 /// Builds the View from model properties
-class SizeTransitionView extends StatefulWidget {
-  final MODEL.SizeTransitionModel model;
+class FadeTransitionView extends StatefulWidget {
+  final MODEL.FadeTransitionModel model;
   final List<Widget> children = [];
   final Widget? child;
   final AnimationController controller;
 
-  SizeTransitionView(this.model, this.child, this.controller)
+  FadeTransitionView(this.model, this.child, this.controller)
       : super(key: ObjectKey(model));
 
   @override
-  SizeTransitionViewState createState() => SizeTransitionViewState();
+  FadeTransitionViewState createState() => FadeTransitionViewState();
 }
 
-class SizeTransitionViewState extends State<SizeTransitionView>
+class FadeTransitionViewState extends State<FadeTransitionView>
     with TickerProviderStateMixin
     implements IModelListener {
   late AnimationController _controller;
@@ -32,6 +32,37 @@ class SizeTransitionViewState extends State<SizeTransitionView>
     super.initState();
 
     _controller = widget.controller;
+    // Tween
+    double from = widget.model.from;
+    double to = widget.model.to;
+    double begin = widget.model.begin;
+    double end = widget.model.end;
+    Curve curve = AnimationHelper.getCurve(widget.model.curve);
+
+    // we must check from != to and begin !< end
+
+    if (begin != 0.0 || end != 1.0) {
+      _animation = Tween<double>(
+        begin: from,
+        end: to,
+      ).animate(CurvedAnimation(
+        curve: new Interval(
+          begin,
+          end,
+          // the style curve to pass.
+          curve: curve,
+        ),
+        parent: _controller,
+      ));
+    } else {
+      _animation = Tween<double>(
+        begin: from,
+        end: to,
+      ).animate(CurvedAnimation(
+        parent: _controller,
+        curve: curve,
+      ));
+    }
   }
 
   @override
@@ -43,7 +74,7 @@ class SizeTransitionViewState extends State<SizeTransitionView>
   }
 
   @override
-  void didUpdateWidget(SizeTransitionView oldWidget) {
+  void didUpdateWidget(FadeTransitionView oldWidget) {
     super.didUpdateWidget(oldWidget);
     if ((oldWidget.model != widget.model)) {
       // re-register model listeners
@@ -71,55 +102,11 @@ class SizeTransitionViewState extends State<SizeTransitionView>
   }
 
   Widget builder(BuildContext context, BoxConstraints constraints) {
-    // Tween
-    double from = widget.model.from;
-    double to = widget.model.to;
-    double begin = widget.model.begin;
-    double end = widget.model.end;
-    Curve curve = AnimationHelper.getCurve(widget.model.curve);
-    //
-    Axis _direction = widget.model.size?.toLowerCase() == "height"
-        ? Axis.vertical
-        : Axis.horizontal;
-    //start, end, center
-    double _align = widget.model.align?.toLowerCase() == "start"
-        ? -1
-        : widget.model.align?.toLowerCase() == "end"
-            ? 1
-            : 0;
-
-    // we must check from != to and begin !< end
-
-    if (begin != 0.0 || end != 1.0) {
-      _animation = Tween<double>(
-        begin: from,
-        end: to,
-      ).animate(CurvedAnimation(
-        curve: new Interval(
-          begin,
-          end,
-          // the style curve to pass.
-          curve: curve,
-        ),
-        parent: _controller,
-      ));
-    } else {
-      _animation = Tween<double>(
-        begin: from,
-        end: to,
-      ).animate(CurvedAnimation(
-        parent: _controller,
-        curve: curve,
-      ));
-    }
-
     // Build View
     Widget? view;
 
-    view = SizeTransition(
-      sizeFactor: _animation,
-      axis: _direction,
-      axisAlignment: _align,
+    view = FadeTransition(
+      opacity: _animation,
       child: widget.child,
     );
