+++ lib/widgets/animation/animation_child/rotate/rotate_transition_view.dart
@@ -1,7 +1,7 @@
 // © COPYRIGHT 2022 APPDADDY SOFTWARE SOLUTIONS INC. ALL RIGHTS RESERVED.
 import 'package:flutter/material.dart';
 import 'package:fml/widgets/animation/animation_helper.dart';
-import 'package:fml/widgets/animation/animation_transition/rotate/rotate_transition_model.dart'
+import 'package:fml/widgets/animation/animation_child/rotate/rotate_transition_model.dart'
     as MODEL;
 import 'package:fml/widgets/widget/widget_model.dart';
