+++ lib/widgets/animation/animation_child/scale/scale_transition_view.dart
@@ -1,27 +1,27 @@
 // © COPYRIGHT 2022 APPDADDY SOFTWARE SOLUTIONS INC. ALL RIGHTS RESERVED.
 import 'package:flutter/material.dart';
 import 'package:fml/widgets/animation/animation_helper.dart';
-import 'package:fml/widgets/animation/animation_transition/fade/fade_transition_model.dart'
+import 'package:fml/widgets/animation/animation_child/scale/scale_transition_model.dart'
     as MODEL;
 import 'package:fml/widgets/widget/widget_model.dart';
 
 /// Animation View
 ///
 /// Builds the View from model properties
-class FadeTransitionView extends StatefulWidget {
-  final MODEL.FadeTransitionModel model;
+class ScaleTransitionView extends StatefulWidget {
+  final MODEL.ScaleTransitionModel model;
   final List<Widget> children = [];
   final Widget? child;
   final AnimationController controller;
 
-  FadeTransitionView(this.model, this.child, this.controller)
+  ScaleTransitionView(this.model, this.child, this.controller)
       : super(key: ObjectKey(model));
 
   @override
-  FadeTransitionViewState createState() => FadeTransitionViewState();
+  ScaleTransitionViewState createState() => ScaleTransitionViewState();
 }
 
-class FadeTransitionViewState extends State<FadeTransitionView>
+class ScaleTransitionViewState extends State<ScaleTransitionView>
     with TickerProviderStateMixin
     implements IModelListener {
   late AnimationController _controller;
@@ -32,37 +32,6 @@ class FadeTransitionViewState extends State<FadeTransitionView>
     super.initState();
 
     _controller = widget.controller;
-    // Tween
-    double from = widget.model.from;
-    double to = widget.model.to;
-    double begin = widget.model.begin;
-    double end = widget.model.end;
-    Curve curve = AnimationHelper.getCurve(widget.model.curve);
-
-    // we must check from != to and begin !< end
-
-    if (begin != 0.0 || end != 1.0) {
-      _animation = Tween<double>(
-        begin: from,
-        end: to,
-      ).animate(CurvedAnimation(
-        curve: new Interval(
-          begin,
-          end,
-          // the style curve to pass.
-          curve: curve,
-        ),
-        parent: _controller,
-      ));
-    } else {
-      _animation = Tween<double>(
-        begin: from,
-        end: to,
-      ).animate(CurvedAnimation(
-        parent: _controller,
-        curve: curve,
-      ));
-    }
   }
 
   @override
@@ -74,7 +43,7 @@ class FadeTransitionViewState extends State<FadeTransitionView>
   }
 
   @override
-  void didUpdateWidget(FadeTransitionView oldWidget) {
+  void didUpdateWidget(ScaleTransitionView oldWidget) {
     super.didUpdateWidget(oldWidget);
     if ((oldWidget.model != widget.model)) {
       // re-register model listeners
@@ -102,11 +71,48 @@ class FadeTransitionViewState extends State<FadeTransitionView>
   }
 
   Widget builder(BuildContext context, BoxConstraints constraints) {
+    // Tween
+    double from = widget.model.from;
+    double to = widget.model.to;
+    double begin = widget.model.begin;
+    double end = widget.model.end;
+    Curve curve = AnimationHelper.getCurve(widget.model.curve);
+
+    //start, end, center
+    Alignment _align =
+        AnimationHelper.getAlignment(widget.model.align?.toLowerCase());
+
+    // we must check from != to and begin !< end
+
+    if (begin != 0.0 || end != 1.0) {
+      _animation = Tween<double>(
+        begin: from,
+        end: to,
+      ).animate(CurvedAnimation(
+        curve: new Interval(
+          begin,
+          end,
+          // the style curve to pass.
+          curve: curve,
+        ),
+        parent: _controller,
+      ));
+    } else {
+      _animation = Tween<double>(
+        begin: from,
+        end: to,
+      ).animate(CurvedAnimation(
+        parent: _controller,
+        curve: curve,
+      ));
+    }
+
     // Build View
     Widget? view;
 
-    view = FadeTransition(
-      opacity: _animation,
+    view = ScaleTransition(
+      alignment: _align,
+      scale: _animation,
       child: widget.child,
     );
