+++ lib/widgets/animation/animation_child/size/size_transition_model.dart
@@ -1,7 +1,7 @@
 // © COPYRIGHT 2022 APPDADDY SOFTWARE SOLUTIONS INC. ALL RIGHTS RESERVED.
 import 'package:flutter/material.dart';
 import 'package:fml/log/manager.dart';
-import 'package:fml/widgets/animation/animation_transition/slide/slide_transition_view.dart';
+import 'package:fml/widgets/animation/animation_child/size/size_transition_view.dart';
 import 'package:fml/widgets/widget/widget_model.dart';
 import 'package:xml/xml.dart';
 import 'package:fml/observable/observable_barrel.dart';
@@ -11,48 +11,46 @@ import 'package:fml/helper/common_helpers.dart';
 
 /// Animation Model
 /// Defines the properties of an [ANIMATION.AnimationView]
-class SlideTransitionModel extends WidgetModel {
+class SizeTransitionModel extends WidgetModel {
   /// Curve starting point from 0.0 to 1.0
-  StringObservable? _from;
+  DoubleObservable? _from;
 
   set from(dynamic v) {
     if (_from != null) {
       _from!.set(v);
     } else if (v != null) {
-      _from = StringObservable(Binding.toKey(id, 'from'), v,
+      _from = DoubleObservable(Binding.toKey(id, 'from'), v,
           scope: scope, listener: onPropertyChange);
     }
   }
 
-  String? get from => _from?.get();
+  double get from {
+    if (_from == null) return 0.0;
+    double f = _from?.get() ?? 0.0;
+    if (f < 0.0) f = 0.0;
+    if (f > 1.0) f = 1.0;
+    return f;
+  }
 
   /// Curve ending point from 1.0 to 0.0
-  StringObservable? _to;
+  DoubleObservable? _to;
 
   set to(dynamic v) {
     if (_to != null) {
       _to!.set(v);
     } else if (v != null) {
-      _to = StringObservable(Binding.toKey(id, 'to'), v,
+      _to = DoubleObservable(Binding.toKey(id, 'to'), v,
           scope: scope, listener: onPropertyChange);
     }
   }
 
-  String get to => _to?.get() ?? "0, 0";
-
-  /// Curve ending point from 1.0 to 0.0
-  StringObservable? _direction;
-
-  set direction(dynamic v) {
-    if (_direction != null) {
-      _direction!.set(v);
-    } else if (v != null) {
-      _direction = StringObservable(Binding.toKey(id, 'direction'), v,
-          scope: scope, listener: onPropertyChange);
-    }
+  double get to {
+    if (_to == null) return 1.0;
+    double f = _to?.get() ?? 1.0;
+    if (f < 0.0) f = 0.0;
+    if (f > 1.0) f = 1.0;
+    return f;
   }
-
-  String? get direction => _direction?.get();
 
   /// Curve
   StringObservable? _curve;
@@ -107,13 +105,41 @@ class SlideTransitionModel extends WidgetModel {
     return f;
   }
 
-  SlideTransitionModel(WidgetModel parent, String? id)
+  /// Curve ending point from 1.0 to 0.0
+  StringObservable? _size;
+
+  set size(dynamic v) {
+    if (_size != null) {
+      _size!.set(v);
+    } else if (v != null) {
+      _size = StringObservable(Binding.toKey(id, 'size'), v,
+          scope: scope, listener: onPropertyChange);
+    }
+  }
+
+  String? get size => _size?.get();
+
+  /// Curve ending point from 1.0 to 0.0
+  StringObservable? _align;
+
+  set align(dynamic v) {
+    if (_align != null) {
+      _align!.set(v);
+    } else if (v != null) {
+      _align = StringObservable(Binding.toKey(id, 'align'), v,
+          scope: scope, listener: onPropertyChange);
+    }
+  }
+
+  String? get align => _align?.get();
+
+  SizeTransitionModel(WidgetModel parent, String? id)
       : super(parent, id); // ; {key: value}
 
-  static SlideTransitionModel? fromXml(WidgetModel parent, XmlElement xml) {
-    SlideTransitionModel? model;
+  static SizeTransitionModel? fromXml(WidgetModel parent, XmlElement xml) {
+    SizeTransitionModel? model;
     try {
-      model = SlideTransitionModel(parent, Xml.get(node: xml, tag: 'id'));
+      model = SizeTransitionModel(parent, Xml.get(node: xml, tag: 'id'));
       model.deserialize(xml);
     } catch (e) {
       Log().debug(e.toString());
@@ -133,7 +159,8 @@ class SlideTransitionModel extends WidgetModel {
     curve = Xml.get(node: xml, tag: 'curve');
     begin = Xml.get(node: xml, tag: 'begin');
     end = Xml.get(node: xml, tag: 'end');
-    direction = Xml.get(node: xml, tag: 'direction');
+    align = Xml.get(node: xml, tag: 'align');
+    size = Xml.get(node: xml, tag: 'size');
   }
 
   @override
@@ -143,6 +170,6 @@ class SlideTransitionModel extends WidgetModel {
   }
 
   Widget getTransitionView(Widget child, AnimationController controller) {
-    return SlideTransitionView(this, child, controller);
+    return SizeTransitionView(this, child, controller);
   }
 }
