+++ lib/widgets/animation/animation_child/size/size_transition_view.dart
@@ -1,35 +1,31 @@
 // © COPYRIGHT 2022 APPDADDY SOFTWARE SOLUTIONS INC. ALL RIGHTS RESERVED.
 import 'package:flutter/material.dart';
-import 'package:fml/helper/string.dart';
 import 'package:fml/widgets/animation/animation_helper.dart';
-import 'package:fml/widgets/animation/animation_transition/slide/slide_transition_model.dart'
+import 'package:fml/widgets/animation/animation_child/size/size_transition_model.dart'
     as MODEL;
 import 'package:fml/widgets/widget/widget_model.dart';
 
 /// Animation View
 ///
 /// Builds the View from model properties
-class SlideTransitionView extends StatefulWidget {
-  final MODEL.SlideTransitionModel model;
+class SizeTransitionView extends StatefulWidget {
+  final MODEL.SizeTransitionModel model;
   final List<Widget> children = [];
   final Widget? child;
   final AnimationController controller;
 
-  SlideTransitionView(this.model, this.child, this.controller)
+  SizeTransitionView(this.model, this.child, this.controller)
       : super(key: ObjectKey(model));
 
   @override
-  FadeTransitionViewState createState() => FadeTransitionViewState();
+  SizeTransitionViewState createState() => SizeTransitionViewState();
 }
 
-class FadeTransitionViewState extends State<SlideTransitionView>
+class SizeTransitionViewState extends State<SizeTransitionView>
     with TickerProviderStateMixin
     implements IModelListener {
   late AnimationController _controller;
-  List<double> _defaultFrom = [-1, 0];
-  late Animation<Offset> _animation;
-  String? _direction;
-  TextDirection? _align;
+  late Animation<double> _animation;
 
   @override
   void initState() {
@@ -47,7 +43,7 @@ class FadeTransitionViewState extends State<SlideTransitionView>
   }
 
   @override
-  void didUpdateWidget(SlideTransitionView oldWidget) {
+  void didUpdateWidget(SizeTransitionView oldWidget) {
     super.didUpdateWidget(oldWidget);
     if ((oldWidget.model != widget.model)) {
       // re-register model listeners
@@ -75,37 +71,29 @@ class FadeTransitionViewState extends State<SlideTransitionView>
   }
 
   Widget builder(BuildContext context, BoxConstraints constraints) {
-    _direction = widget.model.direction?.toLowerCase();
-
-    if (_direction == "right") {
-      _align = TextDirection.ltr;
-      _defaultFrom = [-1, 0];
-    } else if (_direction == "left") {
-      _align = TextDirection.rtl;
-      _defaultFrom = [-1, 0];
-    } else if (_direction == "up") {
-      _defaultFrom = [0, 1];
-    } else if (_direction == "down") {
-      _defaultFrom = [0, -1];
-    }
-
     // Tween
-    List<String>? from = widget.model.from?.split(",");
-    Offset fromOffset = Offset(
-        S.toDouble(from?.elementAt(0)) ?? _defaultFrom[0],
-        S.toDouble(from?.elementAt(1)) ?? _defaultFrom[1]);
-    List<String>? to = widget.model.to.split(",");
-    Offset toOffset = Offset(
-        S.toDouble(to.elementAt(0)) ?? 0, S.toDouble(to.elementAt(1)) ?? 0);
+    double from = widget.model.from;
+    double to = widget.model.to;
     double begin = widget.model.begin;
     double end = widget.model.end;
     Curve curve = AnimationHelper.getCurve(widget.model.curve);
+    //
+    Axis _direction = widget.model.size?.toLowerCase() == "height"
+        ? Axis.vertical
+        : Axis.horizontal;
+    //start, end, center
+    double _align = widget.model.align?.toLowerCase() == "start"
+        ? -1
+        : widget.model.align?.toLowerCase() == "end"
+            ? 1
+            : 0;
+
     // we must check from != to and begin !< end
 
     if (begin != 0.0 || end != 1.0) {
-      _animation = Tween<Offset>(
-        begin: fromOffset,
-        end: toOffset,
+      _animation = Tween<double>(
+        begin: from,
+        end: to,
       ).animate(CurvedAnimation(
         curve: new Interval(
           begin,
@@ -116,9 +104,9 @@ class FadeTransitionViewState extends State<SlideTransitionView>
         parent: _controller,
       ));
     } else {
-      _animation = Tween<Offset>(
-        begin: fromOffset,
-        end: toOffset,
+      _animation = Tween<double>(
+        begin: from,
+        end: to,
       ).animate(CurvedAnimation(
         parent: _controller,
         curve: curve,
@@ -128,9 +116,10 @@ class FadeTransitionViewState extends State<SlideTransitionView>
     // Build View
     Widget? view;
 
-    view = SlideTransition(
-      position: _animation,
-      textDirection: _align,
+    view = SizeTransition(
+      sizeFactor: _animation,
+      axis: _direction,
+      axisAlignment: _align,
       child: widget.child,
     );
