+++ lib/widgets/animation/animation_child/slide/slide_transition_model.dart
@@ -1,54 +1,58 @@
 // © COPYRIGHT 2022 APPDADDY SOFTWARE SOLUTIONS INC. ALL RIGHTS RESERVED.
 import 'package:flutter/material.dart';
 import 'package:fml/log/manager.dart';
-import 'package:fml/widgets/animation/animation_transition/fade/fade_transition_view.dart';
+import 'package:fml/widgets/animation/animation_child/slide/slide_transition_view.dart';
 import 'package:fml/widgets/widget/widget_model.dart';
 import 'package:xml/xml.dart';
 import 'package:fml/observable/observable_barrel.dart';
 import 'package:fml/helper/common_helpers.dart';
 
+/// Progression Curve of an Animation or Transition types
+
 /// Animation Model
 /// Defines the properties of an [ANIMATION.AnimationView]
-class FadeTransitionModel extends WidgetModel {
+class SlideTransitionModel extends WidgetModel {
   /// Curve starting point from 0.0 to 1.0
-  DoubleObservable? _from;
+  StringObservable? _from;
 
   set from(dynamic v) {
     if (_from != null) {
       _from!.set(v);
     } else if (v != null) {
-      _from = DoubleObservable(Binding.toKey(id, 'from'), v,
+      _from = StringObservable(Binding.toKey(id, 'from'), v,
           scope: scope, listener: onPropertyChange);
     }
   }
 
-  double get from {
-    if (_from == null) return 0.0;
-    double f = _from?.get() ?? 0.0;
-    if (f < 0.0) f = 0.0;
-    if (f > 1.0) f = 1.0;
-    return f;
-  }
+  String? get from => _from?.get();
 
   /// Curve ending point from 1.0 to 0.0
-  DoubleObservable? _to;
+  StringObservable? _to;
 
   set to(dynamic v) {
     if (_to != null) {
       _to!.set(v);
     } else if (v != null) {
-      _to = DoubleObservable(Binding.toKey(id, 'to'), v,
+      _to = StringObservable(Binding.toKey(id, 'to'), v,
           scope: scope, listener: onPropertyChange);
     }
   }
 
-  double get to {
-    if (_to == null) return 1.0;
-    double f = _to?.get() ?? 1.0;
-    if (f < 0.0) f = 0.0;
-    if (f > 1.0) f = 1.0;
-    return f;
+  String get to => _to?.get() ?? "0, 0";
+
+  /// Curve ending point from 1.0 to 0.0
+  StringObservable? _direction;
+
+  set direction(dynamic v) {
+    if (_direction != null) {
+      _direction!.set(v);
+    } else if (v != null) {
+      _direction = StringObservable(Binding.toKey(id, 'direction'), v,
+          scope: scope, listener: onPropertyChange);
+    }
   }
+
+  String? get direction => _direction?.get();
 
   /// Curve
   StringObservable? _curve;
@@ -103,13 +107,13 @@ class FadeTransitionModel extends WidgetModel {
     return f;
   }
 
-  FadeTransitionModel(WidgetModel parent, String? id)
+  SlideTransitionModel(WidgetModel parent, String? id)
       : super(parent, id); // ; {key: value}
 
-  static FadeTransitionModel? fromXml(WidgetModel parent, XmlElement xml) {
-    FadeTransitionModel? model;
+  static SlideTransitionModel? fromXml(WidgetModel parent, XmlElement xml) {
+    SlideTransitionModel? model;
     try {
-      model = FadeTransitionModel(parent, Xml.get(node: xml, tag: 'id'));
+      model = SlideTransitionModel(parent, Xml.get(node: xml, tag: 'id'));
       model.deserialize(xml);
     } catch (e) {
       Log().debug(e.toString());
@@ -129,6 +133,7 @@ class FadeTransitionModel extends WidgetModel {
     curve = Xml.get(node: xml, tag: 'curve');
     begin = Xml.get(node: xml, tag: 'begin');
     end = Xml.get(node: xml, tag: 'end');
+    direction = Xml.get(node: xml, tag: 'direction');
   }
 
   @override
@@ -138,6 +143,6 @@ class FadeTransitionModel extends WidgetModel {
   }
 
   Widget getTransitionView(Widget child, AnimationController controller) {
-    return FadeTransitionView(this, child, controller);
+    return SlideTransitionView(this, child, controller);
   }
 }
