+++ lib/widgets/animation/animation_child/slide/slide_transition_view.dart
@@ -1,32 +1,35 @@
 // © COPYRIGHT 2022 APPDADDY SOFTWARE SOLUTIONS INC. ALL RIGHTS RESERVED.
 import 'package:flutter/material.dart';
-import 'package:fml/helper/alignment.dart';
+import 'package:fml/helper/string.dart';
 import 'package:fml/widgets/animation/animation_helper.dart';
-import 'package:fml/widgets/animation/animation_transition/scale/scale_transition_model.dart'
+import 'package:fml/widgets/animation/animation_child/slide/slide_transition_model.dart'
     as MODEL;
 import 'package:fml/widgets/widget/widget_model.dart';
 
 /// Animation View
 ///
 /// Builds the View from model properties
-class ScaleTransitionView extends StatefulWidget {
-  final MODEL.ScaleTransitionModel model;
+class SlideTransitionView extends StatefulWidget {
+  final MODEL.SlideTransitionModel model;
   final List<Widget> children = [];
   final Widget? child;
   final AnimationController controller;
 
-  ScaleTransitionView(this.model, this.child, this.controller)
+  SlideTransitionView(this.model, this.child, this.controller)
       : super(key: ObjectKey(model));
 
   @override
-  ScaleTransitionViewState createState() => ScaleTransitionViewState();
+  FadeTransitionViewState createState() => FadeTransitionViewState();
 }
 
-class ScaleTransitionViewState extends State<ScaleTransitionView>
+class FadeTransitionViewState extends State<SlideTransitionView>
     with TickerProviderStateMixin
     implements IModelListener {
   late AnimationController _controller;
-  late Animation<double> _animation;
+  List<double> _defaultFrom = [-1, 0];
+  late Animation<Offset> _animation;
+  String? _direction;
+  TextDirection? _align;
 
   @override
   void initState() {
@@ -44,7 +47,7 @@ class ScaleTransitionViewState extends State<ScaleTransitionView>
   }
 
   @override
-  void didUpdateWidget(ScaleTransitionView oldWidget) {
+  void didUpdateWidget(SlideTransitionView oldWidget) {
     super.didUpdateWidget(oldWidget);
     if ((oldWidget.model != widget.model)) {
       // re-register model listeners
@@ -72,23 +75,37 @@ class ScaleTransitionViewState extends State<ScaleTransitionView>
   }
 
   Widget builder(BuildContext context, BoxConstraints constraints) {
+    _direction = widget.model.direction?.toLowerCase();
+
+    if (_direction == "right") {
+      _align = TextDirection.ltr;
+      _defaultFrom = [-1, 0];
+    } else if (_direction == "left") {
+      _align = TextDirection.rtl;
+      _defaultFrom = [-1, 0];
+    } else if (_direction == "up") {
+      _defaultFrom = [0, 1];
+    } else if (_direction == "down") {
+      _defaultFrom = [0, -1];
+    }
+
     // Tween
-    double from = widget.model.from;
-    double to = widget.model.to;
+    List<String>? from = widget.model.from?.split(",");
+    Offset fromOffset = Offset(
+        S.toDouble(from?.elementAt(0)) ?? _defaultFrom[0],
+        S.toDouble(from?.elementAt(1)) ?? _defaultFrom[1]);
+    List<String>? to = widget.model.to.split(",");
+    Offset toOffset = Offset(
+        S.toDouble(to.elementAt(0)) ?? 0, S.toDouble(to.elementAt(1)) ?? 0);
     double begin = widget.model.begin;
     double end = widget.model.end;
     Curve curve = AnimationHelper.getCurve(widget.model.curve);
-
-    //start, end, center
-    Alignment _align =
-        AnimationHelper.getAlignment(widget.model.align?.toLowerCase());
-
     // we must check from != to and begin !< end
 
     if (begin != 0.0 || end != 1.0) {
-      _animation = Tween<double>(
-        begin: from,
-        end: to,
+      _animation = Tween<Offset>(
+        begin: fromOffset,
+        end: toOffset,
       ).animate(CurvedAnimation(
         curve: new Interval(
           begin,
@@ -99,9 +116,9 @@ class ScaleTransitionViewState extends State<ScaleTransitionView>
         parent: _controller,
       ));
     } else {
-      _animation = Tween<double>(
-        begin: from,
-        end: to,
+      _animation = Tween<Offset>(
+        begin: fromOffset,
+        end: toOffset,
       ).animate(CurvedAnimation(
         parent: _controller,
         curve: curve,
@@ -111,9 +128,9 @@ class ScaleTransitionViewState extends State<ScaleTransitionView>
     // Build View
     Widget? view;
 
-    view = ScaleTransition(
-      alignment: _align,
-      scale: _animation,
+    view = SlideTransition(
+      position: _animation,
+      textDirection: _align,
       child: widget.child,
     );
