+++ lib/widgets/widget/widget_model.dart
@@ -19,11 +19,11 @@ import 'package:fml/datasources/zebra/model.dart';
 import 'package:fml/system.dart';
 import 'package:fml/widgets/alarm/alarm_model.dart';
 import 'package:fml/widgets/animation/animation_model.dart';
-import 'package:fml/widgets/animation/animation_transition/fade/fade_transition_model.dart';
-import 'package:fml/widgets/animation/animation_transition/rotate/rotate_transition_model.dart';
-import 'package:fml/widgets/animation/animation_transition/scale/scale_transition_model.dart';
-import 'package:fml/widgets/animation/animation_transition/size/size_transition_model.dart';
-import 'package:fml/widgets/animation/animation_transition/slide/slide_transition_model.dart';
+import 'package:fml/widgets/animation/animation_child/fade/fade_transition_model.dart';
+import 'package:fml/widgets/animation/animation_child/rotate/rotate_transition_model.dart';
+import 'package:fml/widgets/animation/animation_child/scale/scale_transition_model.dart';
+import 'package:fml/widgets/animation/animation_child/size/size_transition_model.dart';
+import 'package:fml/widgets/animation/animation_child/slide/slide_transition_model.dart';
 import 'package:fml/widgets/breadcrumb/breadcrumb_model.dart';
 import 'package:fml/widgets/busy/busy_model.dart';
 import 'package:fml/widgets/button/button_model.dart';
